+++ projection/winrt-Microsoft.UI.Xaml/py.Microsoft.UI.Xaml.cpp
@@ -13795,9 +13795,18 @@ namespace py::cpp::Microsoft::UI::Xaml
     {
         try
         {
-            auto value = self->obj.TryLookup(py::convert_to<winrt::Windows::Foundation::IInspectable>(key));
+            auto _key = py::convert_to<winrt::Windows::Foundation::IInspectable>(key);
+            auto value = self->obj.TryLookup(_key);
 
             if (!value) {
+                if constexpr (std::is_base_of_v<winrt::Windows::Foundation::IUnknown, decltype(value)>)
+                {
+                    if (self->obj.HasKey(_key))
+                    {
+                        Py_RETURN_NONE;
+                    }
+                }
+
                 PyErr_SetObject(PyExc_KeyError, key);
                 return nullptr;
             }