+++ projection/winrt-Windows.AI.MachineLearning.Preview/py.Windows.AI.MachineLearning.Preview.cpp
@@ -878,9 +878,18 @@ namespace py::cpp::Windows::AI::MachineLearning::Preview
     {
         try
         {
-            auto value = self->obj.TryLookup(py::convert_to<winrt::hstring>(key));
+            auto _key = py::convert_to<winrt::hstring>(key);
+            auto value = self->obj.TryLookup(_key);
 
             if (!value) {
+                if constexpr (std::is_base_of_v<winrt::Windows::Foundation::IUnknown, decltype(value)>)
+                {
+                    if (self->obj.HasKey(_key))
+                    {
+                        Py_RETURN_NONE;
+                    }
+                }
+
                 PyErr_SetObject(PyExc_KeyError, key);
                 return nullptr;
             }