+++ projection/winrt-Windows.ApplicationModel.Resources.Core/py.Windows.ApplicationModel.Resources.Core.cpp
@@ -2396,9 +2396,18 @@ namespace py::cpp::Windows::ApplicationModel::Resources::Core
     {
         try
         {
-            auto value = self->obj.TryLookup(py::convert_to<winrt::hstring>(key));
+            auto _key = py::convert_to<winrt::hstring>(key);
+            auto value = self->obj.TryLookup(_key);
 
             if (!value) {
+                if constexpr (std::is_base_of_v<winrt::Windows::Foundation::IUnknown, decltype(value)>)
+                {
+                    if (self->obj.HasKey(_key))
+                    {
+                        Py_RETURN_NONE;
+                    }
+                }
+
                 PyErr_SetObject(PyExc_KeyError, key);
                 return nullptr;
             }
@@ -2896,9 +2905,18 @@ namespace py::cpp::Windows::ApplicationModel::Resources::Core
     {
         try
         {
-            auto value = self->obj.TryLookup(py::convert_to<winrt::hstring>(key));
+            auto _key = py::convert_to<winrt::hstring>(key);
+            auto value = self->obj.TryLookup(_key);
 
             if (!value) {
+                if constexpr (std::is_base_of_v<winrt::Windows::Foundation::IUnknown, decltype(value)>)
+                {
+                    if (self->obj.HasKey(_key))
+                    {
+                        Py_RETURN_NONE;
+                    }
+                }
+
                 PyErr_SetObject(PyExc_KeyError, key);
                 return nullptr;
             }
@@ -3562,9 +3580,18 @@ namespace py::cpp::Windows::ApplicationModel::Resources::Core
     {
         try
         {
-            auto value = self->obj.TryLookup(py::convert_to<winrt::hstring>(key));
+            auto _key = py::convert_to<winrt::hstring>(key);
+            auto value = self->obj.TryLookup(_key);
 
             if (!value) {
+                if constexpr (std::is_base_of_v<winrt::Windows::Foundation::IUnknown, decltype(value)>)
+                {
+                    if (self->obj.HasKey(_key))
+                    {
+                        Py_RETURN_NONE;
+                    }
+                }
+
                 PyErr_SetObject(PyExc_KeyError, key);
                 return nullptr;
             }
@@ -3982,9 +4009,18 @@ namespace py::cpp::Windows::ApplicationModel::Resources::Core
     {
         try
         {
-            auto value = self->obj.TryLookup(py::convert_to<winrt::hstring>(key));
+            auto _key = py::convert_to<winrt::hstring>(key);
+            auto value = self->obj.TryLookup(_key);
 
             if (!value) {
+                if constexpr (std::is_base_of_v<winrt::Windows::Foundation::IUnknown, decltype(value)>)
+                {
+                    if (self->obj.HasKey(_key))
+                    {
+                        Py_RETURN_NONE;
+                    }
+                }
+
                 PyErr_SetObject(PyExc_KeyError, key);
                 return nullptr;
             }