+++ projection/winrt-Windows.Media.Core/py.Windows.Media.Core.cpp
@@ -7724,9 +7724,18 @@ namespace py::cpp::Windows::Media::Core
     {
         try
         {
-            auto value = self->obj.TryLookup(py::convert_to<winrt::guid>(key));
+            auto _key = py::convert_to<winrt::guid>(key);
+            auto value = self->obj.TryLookup(_key);
 
             if (!value) {
+                if constexpr (std::is_base_of_v<winrt::Windows::Foundation::IUnknown, decltype(value)>)
+                {
+                    if (self->obj.HasKey(_key))
+                    {
+                        Py_RETURN_NONE;
+                    }
+                }
+
                 PyErr_SetObject(PyExc_KeyError, key);
                 return nullptr;
             }